+++ backend/src/controllers/AuthController.ts
@@ -1,4 +1,4 @@
-import { NextFunction, Request, Response, text } from 'express'
+import type { Request, Response } from 'express'
 import User from '../models/User'
 import { hashPassword } from '../utils/auth'
 import { checkPassword, generateToken } from '../utils/token'
@@ -161,46 +161,8 @@ export class AuthController {
     }
   }
 
-  static user = async (req: Request, res: Response, next: NextFunction): Promise<void> => {
-    const bearer = req.headers.authorization;
-
-    if (!bearer) {
-      res.status(401).json({ error: '認証が必要です' });
-      return;
-    }
-
-    //MEMO: sBearer <JWT> の間のスペースを利用してJWTのみを取得
-    const [, token] = bearer.split(' ');
-
-    if (!token) {
-      res.status(401).json({ error: 'トークンが無効です' });
-      return;
-    }
-
-    try {
-      //MEMO: 秘密鍵とJWTを用いて検証、JWTが有効であればデコードする
-      const decoded = jwt.verify(token, process.env.JWT_SECRET);
-
-      //MEMO: デコードされたトークンがオブジェクトであり、idプロパティを持っていることを確認
-      // if (typeof decoded === 'object' && decoded.id) {
-      //MEMO: 秘密鍵とJWTを用いて検証、JWTが有効であればデコードする
-      //   req.user = await User.findByPk(decoded.id, {
-      //     attributes: ['id', 'name', 'email']
-      //   });
-
-      //   if (!req.user) {
-      //     res.status(401).json({ error: 'ユーザが見つかりません' });
-      //     return;
-      //   }
-
-      //   res.json(req.user); // ユーザー情報をレスポンスとして送信
-      res.json(decoded)
-      // } else {
-      //   res.status(401).json({ error: '無効なトークンです' });
-      // }
-    } catch (error) {
-      res.status(500).json({ error: 'サーバーエラーが発生しました' });
-    }
+  static user = async (req: Request, res: Response): Promise<void> => {
+    res.json(req.user)
   };
 
   static updateCurrentUserPassword = async (
