+++ backend/src/routes/authRouter.ts
@@ -3,6 +3,7 @@ import { AuthController } from '../controllers/AuthController'
 import { handleInputErrors } from '../middleware/validation'
 import { Router } from 'express'
 import { limiter } from '../config/limiter'
+import { authenticate } from '../middleware/auth'
 
 const router = Router()
 
@@ -88,7 +89,11 @@ router.post(
   AuthController.resetPasswordWithToken,
 )
 
-router.get('/user', handleInputErrors, AuthController.user)
+router.get('/user',
+  authenticate,
+  // handleInputErrors,
+  AuthController.user
+)
 
 router.put('/user', handleInputErrors, AuthController.user)
